enviroment pressureEncode

+++ libraries/HouzControl/HouzDevicesCodec.cpp
@@ -44,11 +44,11 @@ deviceData HouzDevicesCodec::decode(String str) { //from serial
 
 //////////////////////////////////////////////////////
 // Pressure patch
-u32 HouzDevicesCodec::pressureEncode(u32 devicePayload){
+u32 HouzDevicesCodec::pressureEncode(float devicePayload){
   return (devicePayload-850)*100; //800hPa offset
 }
 
-u32 HouzDevicesCodec::pressureDecode(u32 devicePayload){
+float HouzDevicesCodec::pressureDecode(u32 devicePayload){
   return (devicePayload/100)+850; //800hPa offset
 }
